+++ models/settings.js
@@ -58,4 +58,6 @@ var SettingsSchema = new Schema({
 	newUsers : Boolean
 });
 
+SettingsSchema.index({ settings_id: 1 }, { unique: true });
+
 module.exports = mongoose.model('Settings', SettingsSchema);